requirements.txt
# Enhanced RAG System Dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP and CORS
python-multipart==0.0.6

# Database and Session Management
# sqlite3-utils==3.35

# Document Processing
pymupdf==1.23.8
python-docx==1.1.0

# Text Processing and NLP
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# HTTP Requests for LLM APIs
requests==2.31.0
httpx==0.25.2

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration Management
python-dotenv==1.0.0

# Security and Authentication
bcrypt==4.1.1
passlib[bcrypt]==1.7.4

# Image Processing (for potential future OCR)
Pillow==10.1.0

# Scientific Computing
numpy==1.24.4

# Utilities
tqdm==4.66.1

# Production Server
gunicorn==21.2.0

# Environment and Logging
python-json-logger==2.0.7

# Type Hints
typing-extensions==4.8.0

# Async Support
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# Development Dependencies (Optional)
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional Security
cryptography==40.0.2


# Performance Monitoring
psutil==5.9.6

# File Type Detection
python-magic==0.4.27

# Unicode Handling
unidecode==1.3.7

# Math and Statistics
scipy==1.11.4

# Date and Time Utilities
python-dateutil==2.8.2

# Cache Support
redis==5.0.1

# Optional: For advanced text processing
spacy==3.7.2

# Optional: For better tokenization
tiktoken==0.5.2

# Optional: For document similarity
scikit-learn==1.3.2

# Optional: For better embedding models
torch==2.1.1
transformers==4.35.2

# Optional: For advanced PDF processing
pdfplumber==0.10.3
